requirements.txt
aws-cdk-lib==2.89.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-go-alpha
